wp-championship v2.6 tag
git-svn-id: https://plugins.svn.wordpress.org/wp-championship/tags/2.6@280023 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ functions.php
@@ -3,12 +3,14 @@
 // 
 // function to show an admin message on an admin page
 //
-function admin_message($msg) {
-  echo "<div class='updated'><p><strong>";
-  echo $msg;
-  echo "</strong></p></div>\n";
+if (!function_exists("admin_message")) 
+{
+    function admin_message($msg) {
+	echo "<div class='updated'><p><strong>";
+	echo $msg;
+	echo "</strong></p></div>\n";
+    }
 }
-
 
 //
 // return an html form field selector for num groups with id
@@ -61,7 +63,7 @@ function calc_points($new=false)
   // for testing
   $new=true;
 
-  // alles zuruekcsetzen
+  // alles zuruecksetzen
   if ($new) {
     $sql = "update $cs_tipp set points=-1 where points <>-1";
     $res = $wpdb->query($sql);
@@ -72,6 +74,11 @@ function calc_points($new=false)
  $cs_pts_tendency=get_option("cs_pts_tendency");    // tendenz
  $cs_pts_supertipp=get_option("cs_pts_supertipp");  // tendenz und tordifferenz
  $cs_pts_champ=get_option("cs_pts_champ");          // championtipp
+ $cs_pts_oneside=get_option("cs_pts_oneside");      // einseitg richtiger tipp 
+ $cs_oneside_tendency=get_option("cs_oneside_tendency");  // einseitg richtiger tipp nur mit richtiger tendenz gültig
+ $cs_goalsum=get_option("cs_goalsum");              // schwellwert für torsumme tipp
+ $cs_pts_goalsum=get_option("cs_pts_goalsum");      // punkte für tosummentipp
+ $cs_goalsum_auto = get_option("cs_goalsum_auto");  // torsummentipp aus tipp berechnen oder separat?
 
  // genauer treffer
  // sql vor mysql 5.0
@@ -92,14 +99,37 @@ function calc_points($new=false)
  //$sql= "update $cs_tipp b set points=1 where mid in ( select a.mid from $cs_match a where a.mid=b.mid and a.result1 <> -1 and a.result2 <> -1 and ( (a.result1<a.result2 and b.result1<b.result2) or (a.result1=a.result2 and b.result1=b.result2) or (a.result1>a.result2 and b.result1>b.result2)  )) and b.points = -1;";
  // mysql 4.x
  $sql= "update $cs_tipp b inner join $cs_match a on a.mid=b.mid and a.result1 <> -1 and a.result2 <> -1 and ( (a.result1<a.result2 and b.result1<b.result2) or (a.result1=a.result2 and b.result1=b.result2) or (a.result1>a.result2 and b.result1>b.result2)  ) set points= $cs_pts_tendency where b.points = -1 and b.result1>-1 and b.result2>-1;";
- $res = $wpdb->query($sql);
- 
+ $res = $wpdb->query($sql); 
+
+ // einseitig richtiger tipp ohne und mit tendenz
+ // mysql 4.x
+ if ($cs_pts_oneside > 0) {
+     if ($cs_oneside_tendency > 0)
+	 $sql= "update $cs_tipp b inner join $cs_match a on a.mid=b.mid and a.result1 <> -1 and a.result2 <> -1 and ( (a.result1<a.result2 and b.result1<b.result2) or (a.result1=a.result2 and b.result1=b.result2) or (a.result1>a.result2 and b.result1>b.result2)  ) and ( a.result1=b.result1 or a.result2=b.result2 ) set points= points + $cs_pts_oneside where b.points = $cs_pts_tendency and b.result1>-1 and b.result2>-1;";  
+     else
+	 $sql= "update $cs_tipp b inner join $cs_match a on a.mid=b.mid and a.result1 <> -1 and a.result2 <> -1 and ( a.result1=b.result1 or a.result2=b.result2 ) set points= $cs_pts_oneside where b.points = -1 and b.result1>-1 and b.result2>-1;";
+     $res = $wpdb->query($sql);
+ }
+
+
  // falscher tipp (setzt alle restlichen auf 0)
  // mysql 5.0
  //$sql= "update  $cs_tipp b set points=0 where mid in ( select a.mid from  a where a.mid=b.mid and a.result1 <> -1 and a.result2 <> -1 ) and b.points = -1;";
  // mysql 4.x
  $sql= "update $cs_tipp b inner join $cs_match a on  a.mid=b.mid and a.result1 <> -1 and a.result2 <> -1  set points=0 where b.points = -1 and b.result1>-1 and b.result2>-1;";
  $res = $wpdb->query($sql);
+ 
+ 
+ // torsummen tipp prüfen und ggf addieren
+ if ($cs_goalsum > 0) {
+     if ($cs_goalsum_auto==0)
+	 $sql= "update $cs_tipp b inner join $cs_match a on a.mid=b.mid and a.result1 <> -1 and a.result2 <> -1 and ( a.result1+a.result2 <= b.result3 and a.result1+a.result2 > $cs_goalsum ) set points=points+$cs_pts_goalsum where b.result1>-1 and b.result2>-1;";
+     else
+	 $sql= "update $cs_tipp b inner join $cs_match a on a.mid=b.mid and a.result1 <> -1 and a.result2 <> -1 and ( a.result1+a.result2 <= b.result1+b.result2 and a.result1+a.result2 > $cs_goalsum ) set points=points+$cs_pts_goalsum where b.result1>-1 and b.result2>-1;"; 
+     $res = $wpdb->query($sql);
+ }
+ 
+
  
  // champion tipp addieren auf finalspiel
  //$sql="update  $cs_tipp b inner join $cs_match a on b.mid=a.mid and a.result1<> -1 and a.result2 <> -1 and a.matchtime = max(a.matchtime) and a.round='F' set points = points +10";
@@ -121,8 +151,9 @@ function calc_points($new=false)
      $sql="update $cs_tipp set points=points + $cs_pts_champ where userid=".$r->userid." limit 1;";
      $wpdb->query($sql);
    }
-   
  }
+ 
+
 }
 
 
@@ -132,7 +163,8 @@ function get_ranking() {
   global $wpdb;
 
   //select fuer ranking der tipper
-  $sql = "select b.user_nicename, a.userid,sum(a.points) as points from $cs_tipp a inner join $wp_users b on a.userid=b.ID where points <> -1 group by b.user_nicename, a.userid order by points DESC;";
+  //$sql = "select b.user_nicename, a.userid,sum(a.points) as points from $cs_tipp a inner join $wp_users b on a.userid=b.ID where points <> -1 group by b.user_nicename, a.userid order by points DESC;"; 
+  $sql = "select b.user_nicename, a.userid,sum(a.points) as points, c.rang as oldrank from $cs_tipp a inner join $wp_users b on a.userid=b.ID inner join $cs_users c on a.userid=c.userid where points <> -1 group by b.user_nicename, a.userid order by points DESC;";
   $res = $wpdb->get_results($sql);
 
   return $res;
@@ -144,21 +176,20 @@ function get_ranking() {
 // zuruekcgeliefert. ist count  = 0 werden alle teams zuruckgegegben
 function get_team_clification($groupid='', $count=0)
 {
-  include("globals.php");
-  global $wpdb;
-
-  //$wpdb->show_errors(true);
-
-  // turniermodus lesen
-  $cs_modus=get_option("cs_modus");
-  
-  // punktvergabe fuer match einlesen
-  $cs_pts_winner=get_option("cs_pts_winner");
-  $cs_pts_looser=get_option("cs_pts_looser");
-  $cs_pts_deuce=get_option("cs_pts_deuce");
-
-  $sql1= <<<EOD
-   create temporary table if not exists cs_tt
+    include("globals.php");
+    global $wpdb;
+    
+    //$wpdb->show_errors(true);
+    
+    // turniermodus lesen
+    $cs_modus      = get_option("cs_modus");
+    // punktvergabe fuer match einlesen
+    $cs_pts_winner = get_option("cs_pts_winner");
+    $cs_pts_looser = get_option("cs_pts_looser");
+    $cs_pts_deuce  = get_option("cs_pts_deuce");
+    
+    $sql1= <<<EOD
+	create temporary table if not exists cs_tt
          select groupid,name,tid,icon,qualified,
 	 sum(result1) as tore,
 	 sum(result2) as gegentore, 
@@ -166,9 +197,9 @@ function get_team_clification($groupid='', $count=0)
 	 from $cs_match 
 	 inner join $cs_team 
 	 on tid=tid1
-	 where winner <>-1 and tid1<>0 and round='V'
+	 where winner<>-1 and tid1<>0 and round='V'
          group by groupid,name,icon,qualified
-	 UNION 
+	 UNION ALL
          select groupid,name,tid,icon,qualified,
 	 sum(result2) as tore,
 	 sum(result1) as gegentore, 
@@ -178,12 +209,12 @@ function get_team_clification($groupid='', $count=0)
 	 on tid=tid2
 	 where winner <>-1 and tid2<>0 and round='V'
          group by groupid,name,icon,qualified
-	 UNION
+	 UNION ALL
          select distinct groupid,name,tid,icon,qualified, 
          0 as tore,0 as gegentore,0 as points
 	 from $cs_match inner join $cs_team on tid=tid1
 	 where winner =-1 and tid1<>0 and round ='V'
-	 UNION
+	 UNION ALL
          select distinct groupid,name,tid,icon, qualified,
          0 as tore,0 as gegentore,0 as points
 	 from $cs_match inner join $cs_team on tid=tid2
@@ -252,7 +283,7 @@ function compare_direct($team1,$team2)
 
   $tore1=0;$tore2=0; $atore1=0; $atore2=0; $winner=-1;
 
-  $wpdb->show_errors(true);
+  //$wpdb->show_errors(true);
 
   $sql="select * from $cs_match where (tid1=$team1 and tid2=$team2) or (tid1=$team2 and tid2=$team1) and winner <> -1";
   $res = $wpdb->get_results($sql);
@@ -390,6 +421,7 @@ EOD;
    }
  }
 
+
  // aktualisiere daten fuer k.o.-runde
  // selektiere noch zu ersetzende pseudo teams 
  $sql0=<<<EOD
@@ -398,13 +430,13 @@ EOD;
    substring( b.name, 2,1 ) as wl,substring(b.name,3) as wlmid,a.matchtime
    FROM $cs_match a
    INNER JOIN $cs_team b ON a.tid1 = b.tid
-   WHERE a.round = 'F' AND (b.name LIKE '#W%' or b.name like '#L%')
+   WHERE a.round = 'F' AND (b.name LIKE '#W%' or b.name like '#V%')
    UNION
    SELECT a.mid,'2' as tnr,a.tid2 as tid,
    substring( b.name, 2,1 ) as wl, substring(b.name,3) as wlmid,a.matchtime
    FROM $cs_match a
    INNER JOIN $cs_team b ON a.tid2 = b.tid
-   WHERE a.round = 'F' AND (b.name LIKE '#W%' or b.name like '#L%')
+   WHERE a.round = 'F' AND (b.name LIKE '#W%' or b.name like '#V%')
    ;
 EOD;
 
@@ -423,9 +455,9 @@ EOD;
    // nur updaten wenn ergebnis vorliegt
    if ( !empty($row) ) {
      // ermittle einzutragendes team
-     if (( $res->wl=='W' and $row->winner==1) or ( $res->wl=='L' and $row->winner==2 ))
+     if (( $res->wl=='W' and $row->winner==1) or ( $res->wl=='V' and $row->winner==2 ))
        $newtid = $row->tid1;
-     if (( $res->wl=='W' and $row->winner==2) or ( $res->wl=='L' and $row->winner==1 ))
+     if (( $res->wl=='W' and $row->winner==2) or ( $res->wl=='V' and $row->winner==1 ))
        $newtid = $row->tid2;
 
      $sql1="update $cs_match set tid".$res->tnr."=".$newtid." where mid=".$res->mid;
@@ -434,7 +466,6 @@ EOD;
    }
  }
 }
-
 
 // verschickt an die abonnenten das aktuelle ranking
 function mailservice()
@@ -450,13 +481,16 @@ function mailservice()
   // ausgabe des aktuellen punktestandes und des ranges
   $rank = get_ranking();
   $i=0;
-  $msg  = "<h2>".__("EM-Tippspiel Mailservice","wpcs")."</h2>\n";
+  $msg  = "<h2>".__("Tippspiel Mailservice","wpcs")."</h2>\n";
   $msg .= "<h2>".__("Aktueller Punktestand","wpcs")."</h2>\n";
   $msg .= "<table border='1' width='500px' cellpadding='0'><thead><tr>\n";
   $msg .= '<th scope="col" style="text-align: center">'.__("Platz","wpcs").'</th>'."\n";
   $msg .= '<th scope="col" style="text-align: center">'.__("Spieler","wpcs").'</th>'."\n";
   $msg .= '<th width="20">'.__("Punktestand","wpcs").'</th>'."\n";
-
+  if (get_option('cs_rank_trend'))
+      $msg .= '<th width="20">'.__("Trend","wpcs").'</th>';
+  $msg .= '</tr></thead>';
+  
   $pointsbefore= -1; 
   $i=0;
   $j=1;
@@ -468,25 +502,108 @@ function mailservice()
     } else
       $j += 1;
     
-    $msg .= "<tr><td align='center'>$i</td><td align='center'>".$row->user_nicename."</td><td align='center'>".$row->points. "</td></tr>";
+    if ($i < $row->oldrank )
+	$trend = "&uArr;";
+    elseif ($i > $row->oldrank )
+	$trend = "&dArr;";
+    else
+	$trend = "&rArr;"; 
     
+    $msg .= "<tr><td align='center'>$i</td><td align='center'>".$row->user_nicename."</td><td align='center'>".$row->points. "</td>";
+    if (get_option('cs_rank_trend'))
+	$msg .= "<td align='center'>$trend</td>";
+    $out .= "</tr>";
+
+
     // gruppenwechsel versorgen
     $pointsbefore = $row->points;
   }
   $msg .= '</table>'."\n<p>&nbsp;";
   
   foreach($res_email as $row) {
-    // mail senden
-    // header bauen
-    $header = "From: webmaster@tuxlog.de";   
-    $header .= "MIME-Version: 1.0\n"; // ohne \r das ist wichtig
-    $header .= "Content-Type: text/html; charset=utf-8\r\n";
+      // mail senden
+      // header bauen
+      $header = "From: " . get_option("admin_email") ."\n"; 
+      $header .= "MIME-Version: 1.0\n"; // ohne \r das ist wichtig
+      $header .= "Content-Type: text/html; charset=utf-8\r\n";
+ 
+      $stat = wp_mail( $row->user_email , "Update Tippspiel" , $msg, $header);
+      if ( $stat) 
+	  echo __("Die email an ","wpcs").$row->user_email.__(" wurde versendet.","wpcs");
+      else 
+	  echo __("Die email an ","wpcs").$row->user_email.__(" konnte <b>nicht</b> versendet werden","wpcs");
+      echo "<br />\n";
+  }
+}
 
-    $stat = mail  ( $row->user_email , "Update EM2008 Tippspiel" , $msg, $header);
-    if ( $stat) 
-      echo __("Die email an ","wpcs").$row->user_email.__(" wurde versendet.<br />","wpcs");
-    else 
-      echo __("Die email an ","wpcs").$row->user_email.__("konnte <b>nicht</b> versendet werden","wpcs");
+// verschickt an die abonnenten erinnerungen falls noch nicht getippt wurde
+function mailservice2()
+{
+    // prüfen ob wir erinnern sollen
+    $cs_reminder= get_option("cs_reminder");
+    if ( ! $cs_reminder)
+	return;
+
+    // globale variable einlesen
+    include("globals.php");
+    global $wpdb;
+  
+
+    // holen der match ids fuer die spiele die noch nicht angefangen haben aber in
+    // den nächsten stunden anfangen
+    $cs_reminder_hours = get_option("cs_reminder_hours");
+    $now  =  time() + ( get_option( 'gmt_offset' ) * 3600 );
+    $then =  $now + ( $cs_reminder_hours * 3600 );
+
+    $mnow  = gmdate( 'Y-m-d H:i:s', $now );
+    $mthen = gmdate( 'Y-m-d H:i:s', $then );
+
+    $sql = "select mid from cs_match where matchtime > '$mnow' and matchtime <= '$then';";
+    $res_mid = $wpdb->get_results($sql);
+
+
+
+    $mids="(";
+    foreach ($res_mid as $row)
+	$mids .= $row->mid . ", ";
+
+    $mids .= "-9999)";
+
+    // holen der userids, die fuer diese match ids noch nicht getippt haben
+    $sql ="select a.userid, b.mid from cs_users a, cs_match b where b.mid in $mids and not exists ( select userid, mid from cs_tipp where userid=a.userid and mid=b.mid ) order by a.userid, b.mid;";
+    $res_user = $wpdb->get_results($sql);
+
+    // fuer jeden user mit fehlendem tipp email zusammenstellen und senden
+    foreach ( $res_user as $u )
+    {
+	// email adresse holen
+	$sql="select user_nicename, user_email from $wp_users where ID=$u->userid;";
+	$res_email=$wpdb->get_results($sql);
+
+	// match daten adresse holen
+	$sql="select b.name name1,c.name name2,a.matchtime,a.location from cs_match a inner join cs_team b on a.tid1 = b.tid inner join cs_team c on a.tid2=c.tid where mid=$u->mid;";
+	$res_match=$wpdb->get_results($sql);
+
+
+	// mailnachricht zusammen bauen
+	$msg  = "<h2>".__("Tippspiel Mailservice","wpcs")."</h2>\n";
+	$msg .= "<h2>".__("Möchtest du noch tippen? Das folgende Spiel beginnt bald: ","wpcs")."</h2>\n";
+	$msg .= $res_match[0]->name1 . " : " . $res_match[0]->name2 . " in " . $res_match[0]->location . " startet " . $res_match[0]->matchtime . "<br /><br/>\n";
+	$msg .= "Viel Glück wünscht der Tippspiel-Admin.\n";
+   
+
+	// header bauen
+	$header = "From: " . get_option("admin_email") . "\n";
+	$header .= "MIME-Version: 1.0\n"; // ohne \r das ist wichtig
+	$header .= "Content-Type: text/html; charset=utf-8\r\n";
+
+      	// mail senden
+	$stat = wp_mail  ( $res_email[0]->user_email , "Update Tippspiel" , $msg, $header);
+	if ( $stat) 
+	    echo __("Die Erinnerungsemail an ","wpcs").$res_email[0]->user_email.__(" wurde versendet.","wpcs");
+	else 
+	    echo __("Die Erinnerungsemail an ","wpcs").$res_email[0]->user_email.__(" konnte <b>nicht</b> versendet werden","wpcs");
+	echo "<br />";
   }
 }
 
@@ -523,7 +640,7 @@ function get_team_stats($teamid)
   include("globals.php");
   global $wpdb;
   
-  $wpdb->show_errors(true);
+  //$wpdb->show_errors(true);
   
   // anzahl spiele
   $sql1= "select count(*) as anz1 from $cs_match where round='V' and (tid1=$teamid or tid2=$teamid) and winner <> -1; ";
@@ -544,5 +661,55 @@ function get_team_stats($teamid)
   
   return array('spiele' => $res1->anz1, 'siege' => $res2->anz2, 
 	       'unentschieden' => $res3->anz3, 'niederlagen' => $res4->anz4 );
+}
+
+
+function get_float_js() {
+$js = <<<EOL
+<script type="text/javascript">
+var name = "#WPCSfloatMenu";
+var menuYloc = null;
+	
+jQuery(document).ready(function(){
+	menuYloc = parseInt(jQuery(name).css("top").substring(0,jQuery(name).css("top").indexOf("px")))
+	    jQuery(window).scroll(function () { 
+		    offset = menuYloc+jQuery(document).scrollTop()+"px";
+		    jQuery(name).animate({top:offset},{duration:500,queue:false});
+		});
+    }); 
+</script>
+EOL;
+
+return $js;
+}
+
+//
+// speichert die aktuelle platzierung alle mitspieler in der tabelle cs_users
+// in der spalte rang
+//
+function store_current_ranking() {
+    include("globals.php");
+    global $wpdb;
+
+    $pointsbefore=-1;
+    $i=0;
+    $j=1;
+    // hole aktuelle mitspielerplatzierung
+    $rank = get_ranking();
+
+    foreach ($rank as $row) {
+	// platzierung erhoehen, wenn punkte sich veraendern
+	if ($row->points != $pointsbefore) {
+	    $i += $j;
+	    $j = 1;
+	} else
+	    $j += 1;
+    
+    	$sql="update $cs_users set rang=$i where userid=$row->userid;";
+	$wpdb->query($sql);
+    
+	// gruppenwechsel versorgen
+	$pointsbefore = $row->points;
+    }
 }
 ?>
